llm-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.8.0
google-api-core>=2.26.0
python-dotenv>=1.0.0
slowapi>=0.1.9
redis>=5.0.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
